taken reviews into consideration
